Refactor tab item drag logic
